% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_string.R
\name{common_string}
\alias{common_string}
\title{common_string}
\usage{
common_string(entity_1, entity_2)
}
\arguments{
\item{entity_1}{A string containing the first word to be compared.}

\item{entity_2}{A string containing the second word to be compared.}
}
\value{
The number of common letters at similar positions.
}
\description{
common_string
}
\examples{
common_string("Paul", "Paul Panther")

common_string("Maria", "Paul")
}
