% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicates.R
\name{find_duplicates}
\alias{find_duplicates}
\title{find_duplicates}
\usage{
find_duplicates(dataframe, prepared_duplicate_data, min_similarity)
}
\arguments{
\item{dataframe}{The raw dataframe where duplicates should be identified.}

\item{prepared_duplicate_data}{A prepared form of the dataframe after using the function "prepare_duplicate_dataframe".}

\item{min_similarity}{A numeric specifying the similarity threshold for two strings to be identified as duplicate.}
}
\value{
Adds a column "duplicate" to the initial dataframe that contains "TRUE" for all duplicates.
}
\description{
find_duplicates
}
\examples{
entity_id <- c(1:4)
entity <- c("Philipp Müller", "Miley Cyrus", "Cyrus", "Wagner")
document_id <- c(1, 1, 1, 1)
actors_in_document_1 <- data.frame(entity_id, entity, document_id)

prepared_dataframe_actors_in_document_1 <- prepare_duplicate_dataframe(actors_in_document_1)

actors_in_document_1_with_duplicates <- find_duplicates(actors_in_document_1, prepared_dataframe_actors_in_document_1, 0.8)
}
