% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_relevant_actors.R
\name{mark_relevant_actors}
\alias{mark_relevant_actors}
\title{Mark relevant actors}
\usage{
mark_relevant_actors(dataframe, prepared_duplicate_data, min_similarity)
}
\arguments{
\item{dataframe}{A dataframe with identified duplicates.}

\item{prepared_duplicate_data}{A prepared form of the dataframe after using the function "prepare_duplicate_dataframe".}

\item{min_similarity}{A numeric specifying the similarity threshold for two strings to be identified as duplicate.}
}
\value{
Changes the values of the column "relevant" in the initial dataframe to distinguish relevant actors from irrelevant actors and duplicates.
}
\description{
Mark relevant actors
}
\examples{
entity_id <- c(1:4)
entity <- c("Philipp Müller", "Miley Cyrus", "Cyrus", "Wagner")
document_id <- c(1, 1, 1, 1)
relevant <- c(TRUE, TRUE, TRUE, FALSE)
duplicate <- c(NA, NA, TRUE, NA)
actors_in_document_1 <- data.frame(entity_id, entity, document_id, relevant, duplicate)

prepared_dataframe_actors_in_document_1 <- prepare_duplicate_dataframe(actors_in_document_1)

relevant_actors_in_document_1 <- mark_relevant_actors(actors_in_document_1, prepared_dataframe_actors_in_document_1, 0.8)
}
