% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_duplicate_dataframe.R
\name{prepare_duplicate_dataframe}
\alias{prepare_duplicate_dataframe}
\title{prepare_duplicate_dataframe}
\usage{
prepare_duplicate_dataframe(dataframe)
}
\arguments{
\item{dataframe}{A dataframe to be transformed.}
}
\value{
A dataframe with 6 columns.
}
\description{
prepare_duplicate_dataframe
}
\examples{
entity_id <- c(1:4)
entity <- c("Philipp Müller", "Miley Cyrus", "Cyrus", "Wagner")
document_id <- c(1, 1, 1, 1)
actors_in_document_1 <- data.frame(entity_id, entity, document_id)

prepared_dataframe_actors_in_document_1 <- prepare_duplicate_dataframe(actors_in_document_1)
}
